+++ projects/CalibrationStatusClassification/deploy/__init__.py
@@ -1,0 +1,21 @@
+"""
+Calibration Status Classification Model Deployment Package
+
+This package provides utilities for exporting, verifying, and evaluating
+CalibrationStatusClassification models in ONNX and TensorRT formats.
+"""
+
+from .config import DeploymentConfig, parse_args, setup_logging
+from .evaluator import evaluate_exported_model, print_evaluation_results
+from .exporters import export_to_onnx, export_to_tensorrt, run_verification
+
+__all__ = [
+    "DeploymentConfig",
+    "parse_args",
+    "setup_logging",
+    "export_to_onnx",
+    "export_to_tensorrt",
+    "run_verification",
+    "evaluate_exported_model",
+    "print_evaluation_results",
+]
